pwaのステータス表示処理をfooterからProviderに移動
